Made operations line based to align with editor API
The change events in ace editor look something like this:

```
{"start":{"row":1,"column":4},"action":"insert","lines":["o"]}
{"start":{"row":1,"column":5},"action":"insert","lines":["",""]}
```

(Simplified)

So, the new libbuf API aligns with this. The 2 main advantages are:

* Easy to implement, frontend doesn't have to translate events
* Concurrent editing on separate lines should just work

There's still more work to do to make the algorithm more robust,
but I think this is good enough for now. It should be better
than the old version that is deployed currently.

Both inserting and removing seems to work well, both in unit tests
and in editor.

Signed-off-by: Ole Herman Schumacher Elgesem <oleherman93@gmail.com>